refactor(admin-spa): extract admin sidebar
